refactor: kod okunabilirliği ve reusable için global css ve react bileşenlerini modüler yapıya geçir
- Global ve yerel CSS mimarisine geçiş yap
- Ortak react bileşenlerini components/ui dizininde toparla
- Sayfalara özel bileşenleri sayfa dosyalarında bırak
